Shift things back into peripheral.generic
I'm not planning to do any mod integration in CC:T, so let's not pretend
we will.
